requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data Processing
scipy>=1.7.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
xarray>=0.20.0

# Visualization Extensions
plotly>=5.0.0
bokeh>=2.4.0
holoviews>=1.14.0
colorcet>=3.0.0
cmasher>=1.6.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0